Fix 41 pre-existing compile errors: add lib target, async_trait, borrow fixes, Utf8Bytes

+++ rust-backend/src/react_loop.rs
@@ -660,7 +660,7 @@ mod tests {
             result.final_answer.is_some(),
             "يجب أن تكون هناك إجابة نهائية"
         );
-        assert_eq!(result.stop_reason as u8, StopReason::Completed as u8);
+        assert_eq!(result.stop_reason, StopReason::Completed);
     }
 }
